+++ web/src/js/ads/openx/openxConfig.js
